# refactor: use cached paths to improve performance