#### Is this adding or improving a _feature_ or fixing a _bug_?

To improve performance.

#### What's the new behavior?

No behavior of API will be changed.

#### How does this change work?

1. Generate paths' cache for each node in a document tree.
1. Use path to locate a node to get/modify it.
    > If the API accept keys, just map key to path with `this.getPath(key)`
1. Update paths' cache if the structure of document tree is changed.
1. The new document tree will inherit updated paths' cache.

After this PR, all the methods that won't modify the structure of document tree will improve from O(n) to O(d), e.g. all the getters and `add_mark` `set_mark` `remove_mark` `set_node` `insert_text` `move_text`. 

And we only need to traverse the sub tree when its structure is changed.

1. n => the number of nodes of a document
1. d => the depth of a node in document
In real world, d << n.

Reviewers: @ianstormtaylor @zhujinxuan