@benjycui 2018-05-04
Both of @zhujinxuan and me agree to map `key` to `path` to improve the performance of get/modify.

The only difference is that we use different way to improve the performance of `getPath`.