@zhujinxuan 2018-05-04
A small question: why use cache as Map() not `{}.__proto__ =null`; I remember, at least before last July, the Map is slower in get when the key is all string.