@zhujinxuan 2018-05-04
Great, we only need to update with `updateNode` and `insertNode` and thus, then we can always have good path.

Only one problem is that shall we recursively update the path?  If we shall, how to recursively set path without losing efficiency?