@zhujinxuan 2018-05-04
```
    changes
      delete-backward
        user : 328.72 → 552.30 ops/sec (68.01259702935722% faster)
        hr : 411.24 → 909.28 ops/sec (121.10671342770662% faster) 🙌
      delete-forward
        user : 111.04 → 164.88 ops/sec (48.47736408392881% faster)
        hr : 163.67 → 271.09 ops/sec (65.63535241091316% faster)
      insert-node-by-key
        user : 223.22 → 626.18 ops/sec (180.5201210854309% faster) 🙌
        hr : 332.78 → 989.49 ops/sec (197.3412813963255% faster) 🙌
      insert-text-by-key-multiple
        user : 69.15 → 165.07 ops/sec (138.70555589487432% faster) 🙌
        hr : 91.77 → 259.41 ops/sec (182.68057222144358% faster) 🙌
      insert-text-by-key
        user : 366.01 → 1422.02 ops/sec (288.51988951531894% faster) 🙌
        hr : 407.11 → 2343.71 ops/sec (475.69453689890497% faster) 🙌
      insert-text
        user : 1685.80 → 2466.68 ops/sec (46.32065235296292% faster)
        hr : 1981.42 → 4206.84 ops/sec (112.31377436845244% faster) 🙌
      normalize
        user : 658.01 → 834.76 ops/sec
        hr : 981.88 → 1283.84 ops/sec
      remove-node-by-key
        user : 286.73 → 788.50 ops/sec (174.9940279541578% faster) 🙌
        hr : 368.83 → 1314.08 ops/sec (256.2861357698833% faster) 🙌
      split-block
        user : 379.03 → 479.18 ops/sec
        hr : 479.96 → 805.65 ops/sec (67.85534893571239% faster)
    models
      from-json-big
        user : 3.61 → 4.21 ops/sec
        hr : 3.89 → 6.02 ops/sec (54.73709720074705% faster)
      from-json
        user : 171.05 → 184.17 ops/sec
        hr : 244.77 → 308.29 ops/sec
      get-active-marks-at-range
        user : 237.30 → 476.64 ops/sec (100.85639437241885% faster) 🙌
        hr : 283.34 → 580.91 ops/sec (105.02146443489802% faster) 🙌
      get-ancestors
        user : 4026.25 → 31138.55 ops/sec (673.388731106515% faster) 🙌
        hr : 5688.91 → 63763.42 ops/sec (1020.8362576537926% faster) 😱
      get-blocks-at-range
        user : 663.70 → 1125.98 ops/sec (69.65123588575764% faster)
        hr : 1085.43 → 2309.03 ops/sec (112.72907185785827% faster) 🙌
      get-blocks
        user : 6288.14 → 8458.17 ops/sec (34.509706892742734% faster)
        hr : 7691.27 → 11956.95 ops/sec (55.46146506613449% faster)
      get-characters-at-range
        user : 223.19 → 287.74 ops/sec
        hr : 212.90 → 346.63 ops/sec (62.814437149237904% faster)
      get-characters
        user : 3576.32 → 3666.84 ops/sec
        hr : 3615.92 → 5178.32 ops/sec (43.20911949710848% faster)
      get-common-ancestor
        user : 693.82 → 29848.73 ops/sec (4202.10177012649% faster) 😱
        hr : 580.76 → 52352.47 ops/sec (8914.438069644988% faster) 😱
      get-furthest-ancestor
        user : 7001.33 → 60471.01 ops/sec (763.7080848820091% faster) 🙌
        hr : 8924.66 → 173701.55 ops/sec (1846.310076945164% faster) 😱
      get-inlines-at-range
        user : 575.41 → 1851.25 ops/sec (221.72900530343247% faster) 🙌
        hr : 689.16 → 2788.99 ops/sec (304.6952019842218% faster) 🙌
      get-inlines
        user : 10796.53 → 10307.06 ops/sec
        hr : 14964.08 → 15498.44 ops/sec
      get-leaves
        user : 8513.67 → 9982.92 ops/sec
        hr : 14370.33 → 16258.30 ops/sec
      get-marks-at-range
        user : 155.80 → 180.06 ops/sec
        hr : 171.91 → 246.56 ops/sec (43.427610027276145% faster)
      get-marks
        user : 507.37 → 496.00 ops/sec
        hr : 787.88 → 870.32 ops/sec
      get-parent
        user : 6092.50 → 34265.24 ops/sec (462.41660627684666% faster) 🙌
        hr : 6893.33 → 66451.77 ops/sec (864.0014752019548% faster) 🙌
      get-path
        user : 2136.64 → 123532.77 ops/sec (5681.646459986303% faster) 😱
        hr : 2850.94 → 309911.71 ops/sec (10770.52453020746% faster) 😱
      get-texts-at-range
        user : 4183.46 → 6814.66 ops/sec (62.89533621168808% faster)
        hr : 5315.87 → 9748.23 ops/sec (83.37990384638904% faster)
      get-texts
        user : 10717.81 → 9654.29 ops/sec
        hr : 15128.62 → 14774.64 ops/sec
      getDescendant
        user : 8863.47 → 45246.80 ops/sec (410.4863865126614% faster) 🙌
        hr : 10989.88 → 105839.97 ops/sec (863.0677839232625% faster) 🙌
      has-node-multiple
        user : 1907.49 → 9454.65 ops/sec (395.6587471357242% faster) 🙌
        hr : 2165.12 → 11884.54 ops/sec (448.9090326390179% faster) 🙌
      has-node
        user : 7527.37 → 41278.57 ops/sec (448.3799985897573% faster) 🙌
        hr : 9577.25 → 93003.25 ops/sec (871.0846629235726% faster) 🙌
      to-json
        user : 827.97 → 677.72 ops/sec
        hr : 1147.06 → 1104.63 ops/sec
      update-node
        user : 2909.93 → 5569.27 ops/sec (91.38808219878119% faster)
        hr : 3483.29 → 10554.17 ops/sec (202.99471083297814% faster) 🙌
    html-serializer
      deserialize
        user : 8.28 → 26.33 ops/sec (217.99988521337679% faster) 🙌
        hr : 10.06 → 35.49 ops/sec (252.8654389926683% faster) 🙌
      serialize
        user : 203.20 → 252.72 ops/sec
        hr : 325.46 → 411.25 ops/sec
    plain-serializer
      deserialize
        user : 405.47 → 420.02 ops/sec
        hr : 647.01 → 684.20 ops/sec
      serialize
        user : 2520.71 → 2379.43 ops/sec
        hr : 3628.95 → 3599.25 ops/sec
    rendering
      normal
        user : 66.02 → 77.60 ops/sec
        hr : 80.76 → 106.41 ops/sec

✨  Done in 61.09s.
```

Crazy improvement