@benjycui 2018-05-05
> A small question: why use cache as Map() not {}.__proto__ =null; I remember, at least before last July, the Map is slower in get when the key is all string.

@zhujinxuan we will reuse paths' cache, an immutable Map means that we don't need to shallow copy the cache. See: https://github.com/ianstormtaylor/slate/pull/1818/files#diff-54c7838b3ea52077be8fa46cee479d9bR2057