@benjycui 2018-05-05
> Only one problem is that shall we recursively update the path? If we shall, how to recursively set path without losing efficiency?

In real world, a document tree will look like this:

![image](https://user-images.githubusercontent.com/3580607/39658627-a4dfab6a-5049-11e8-8c27-95f911c15f1b.png)

1. The depth is small, for users are not going to write a deep nested document which is hard to read. 
    > This is why I said O(d) << O(n)
1. There are many branches in first layer of document tree. So that we only need to traverse a small subtree unless we had insert/remove node to `document.nodes`.
1. Yep, it is a common case to insert/remove node to `document.nodes`(e.g. insert new line, table), but users won't do that frequently. What users may do frequently is `isnert_text` `remove_text` which will not change the document tree.